+++ autodarts-desktop/App.axaml.cs
@@ -30,40 +30,6 @@ namespace autodarts_desktop
         }
         
         
-        /*
-        private const string UniqueEventName = "{AUTODARTS-DESKTOP-STATE-RUNNING}";
-
-        private void App_Startup(object sender, object e)
-        {
-            bool isNewInstance = false;
-
-            using (var mutex = new System.Threading.Mutex(true, UniqueEventName, out isNewInstance))
-            {
-                if (!isNewInstance)
-                {
-                    //Shutdown();
-                    Environment.Exit(0);
-                    return;
-                }
-
-                var current_process = Process.GetCurrentProcess();
-                var other_process = Process.GetProcessesByName(current_process.ProcessName).FirstOrDefault(p => p.Id != current_process.Id);
-
-                if (other_process != null && other_process.MainWindowHandle != IntPtr.Zero)
-                {
-                    if (IsIconic(other_process.MainWindowHandle))
-                    {
-                        ShowWindow(other_process.MainWindowHandle, SW_RESTORE);
-                    }
-                    SetForegroundWindow(other_process.MainWindowHandle);
-                    //Shutdown();
-                    Environment.Exit(0);
-                }
-            }
-        }
-        */
-        
-
 
         [DllImport("user32")]
         static extern bool IsIconic(IntPtr hWnd);